Divide code into more modules
