+++ index.js
@@ -1,11 +1,5 @@
 /* eslint-disable max-classes-per-file */
 /* eslint-disable no-use-before-define */
-import { DateTime } from './modules/luxon.js';
-import Book from './modules/book.js';
-
-const list = document.querySelector('.book-lists');
-const add = document.querySelector('.addnewbook');
-const contact = document.querySelector('.contact');
-const listlink = document.querySelector('.list-link');
-const addlink = document.querySelector('.add-link');
-const contactlink = document.querySelector('.contact-link');
+import './modules/navigation.js';
+import './modules/display.js';
+import './modules/datetime.js';
