+++ modules/navigation.js
@@ -1,3 +1,10 @@
+const list = document.querySelector('.book-lists');
+const add = document.querySelector('.addnewbook');
+const contact = document.querySelector('.contact');
+const listlink = document.querySelector('.list-link');
+const addlink = document.querySelector('.add-link');
+const contactlink = document.querySelector('.contact-link');
+
 listlink.addEventListener('click', () => {
   list.style.display = 'block';
   listlink.style.color = 'blue';
